Fixed compiler warning w.r.t casting a time_t to an int, but I think the better fix would be to use a uint as I think the time could wrap.
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/trunk/microemacs@2022 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ src/winterm.c
@@ -491,7 +491,7 @@ gettimeofday (struct meTimeval *tp, struct meTimezone *tz)
     UNREFERENCED_PARAMETER (tz);
 
     /* Get the second resolution time */
-    tp->tv_sec = time(NULL) ;
+    tp->tv_sec = (int) time(NULL) ;
 
     /* Get the microsecond time */
     GetLocalTime(&stime) ;
